+++ config/routes.rb
@@ -131,6 +131,7 @@ Rails.application.routes.draw do
       resources :lists do
         collection do
           post :generate
+          post :expand
         end
       end
       resources :stratification_factors do